Misc. changes; likely doesnt work. gonna rewrite after merge

+++ kernel/mailbox.c
@@ -4,7 +4,7 @@
 #include "kernel/board.h"
 #include "kernel/mailbox.h"
 
-uint32_t mbox_read(uint8_t channel) {
+uint32_t mbox_read(uint32_t channel) {
 	for (;;) {
 		while ((MMIO_READ(uint32_t, MAIL0_STATUS) & MAIL_EMPTY) != 0);  // Wait for data
 
@@ -22,3 +22,4 @@ void mbox_write(uint8_t channel, uint32_t value) {
 
 	MMIO_WRITE(uint32_t, MAIL1_WRITE, value);
 }
+
